+++ NEWS.md
@@ -1,3 +1,3 @@
 # SpaCE 1.0.0
 
-* Release SpaCE package.
+* Release SpaCET package.